bobashare::storage::file

Struct FileBackend

pub struct FileBackend {
    pub path: PathBuf,
}

A directory on disk which is used to store uploads

Fields
path: PathBuf

path of the directory containing all uploads

Implementations

impl FileBackend

pub async fn new(path: PathBuf) -> Result<Self, NewBackendError>

Construct a file backend, creating the directory if it doesnโt exist.

impl FileBackend

pub async fn create_upload<S: AsRef<str>>( &self, id: S, filename: S, mimetype: Mime, expiry: Option<TimeDelta>, delete_key: Option<String>, ) -> Result<UploadHandle, CreateUploadError>

impl FileBackend

pub async fn read_upload_metadata<S: AsRef<str>>( &self, id: S, ) -> Result<Upload, OpenUploadError>

pub async fn open_upload<S: AsRef<str>>( &self, id: S, write: bool, ) -> Result<UploadHandle, OpenUploadError>

does not check if the upload is expired, do that yourself

impl FileBackend

pub async fn delete_upload<S: AsRef<str>>( &self, id: S, ) -> Result<(), DeleteUploadError>

impl FileBackend

Checks if an upload is valid or if it should be cleaned up

Checks:

check if there is an upload file

check if there is a metadata file

make sure the file isnโt locked

check if the metadata file is valid (meaning: can be deserialized)

migrate the metadata if needed

check if the upload has expired

TODO: what should we do if itโs a file and not a directory?

pub async fn validate_upload<S: AsRef<str>>( &self, id: S, ) -> Result<ValidateResult, ValidateError>

impl FileBackend

pub async fn cleanup(&self) -> Result<(), CleanupError>

Validate all the uploads in the repository and delete ones that are invalid.

See FileBackend::validate_upload for the checks that are performed