Module bobashare::storage::file

A backend where uploads are stored as files on disk

Structs

FileBackend

A directory on disk which is used to store uploads

Enums

CleanupError

Errors when running a repository cleanup task

CreateUploadError

Errors when creating an upload in a file backend

DeleteUploadError

Errors when deleting an upload stored in a file backend

InvalidReason

Reasons why an upload might be invalid and need to be deleted

NewBackendError

Errors when creating a new FileBackend

OpenUploadError

Errors when opening an upload stored in a file backend

ValidateError

Critical errors when validating an upload that mean we can’t determine whether it’s valid or not

ValidateResult